Extract a separate StringExtensions class (win)
